Update RUN THIS FIRST.py
Updated the code to a split-function version by commenting it. Only one of which can be activated or uncommented by the administrator and the task will almost change as whether they choose insignificant time changes (including only hours, minutes and seconds) would be counted to the database or not.
These Insignificant Changes can leave significant tolls over time on performance and quality of the project. (DON'T FORGET TO COMMENT ONE AFTER THE OTHER HAS BEEN UNCOMMENTED OR OTHERWISE IT WOULD END UP WITH AN INTERRUPTION)